+++ src/test/java/kitchenpos/menu/service/ProductServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.menu.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,11 +6,11 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.lang.reflect.InvocationTargetException;
 import java.math.BigDecimal;
 import java.util.List;
+import kitchenpos.exception.InvalidNumberException;
 import kitchenpos.menu.dto.request.CreateProductRequest;
 import kitchenpos.menu.dto.response.ProductResponse;
-import kitchenpos.exception.InvalidNumberException;
-import kitchenpos.menu.service.ProductService;
 import kitchenpos.util.ObjectCreator;
+import kitchenpos.util.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,7 +29,8 @@ class ProductServiceTest extends ServiceTest {
             throws InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         // given
         final int newProductId = productService.list().size() + 1;
-        final CreateProductRequest request = ObjectCreator.getObject(CreateProductRequest.class, "test", BigDecimal.valueOf(100));
+        final CreateProductRequest request = ObjectCreator.getObject(CreateProductRequest.class, "test",
+                BigDecimal.valueOf(100));
 
         // when
         final ProductResponse actual = productService.create(request);
@@ -43,7 +44,8 @@ class ProductServiceTest extends ServiceTest {
     void create_FailPrice()
             throws InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException {
         // given
-        final CreateProductRequest request = ObjectCreator.getObject(CreateProductRequest.class, "test", BigDecimal.valueOf(-1));
+        final CreateProductRequest request = ObjectCreator.getObject(CreateProductRequest.class, "test",
+                BigDecimal.valueOf(-1));
 
         // when
         assertThatThrownBy(() -> productService.create(request))
