+++ src/test/java/kitchenpos/table/service/TableServiceTest.java
@@ -1,19 +1,19 @@
-package kitchenpos.application;
+package kitchenpos.table.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.stream.Stream;
+import kitchenpos.exception.EmptyListException;
+import kitchenpos.exception.InvalidNumberException;
+import kitchenpos.exception.NoSuchDataException;
 import kitchenpos.table.dto.request.ChangeEmptyRequest;
 import kitchenpos.table.dto.request.ChangeNumberOfGuestsRequest;
 import kitchenpos.table.dto.request.CreateOrderTableRequest;
 import kitchenpos.table.dto.response.OrderTableResponse;
-import kitchenpos.exception.EmptyListException;
-import kitchenpos.exception.InvalidNumberException;
-import kitchenpos.exception.NoSuchDataException;
-import kitchenpos.table.service.TableService;
 import kitchenpos.util.ObjectCreator;
+import kitchenpos.util.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
