+++ src/test/java/kitchenpos/table/ui/TableGroupRestControllerTest.java
@@ -1,21 +1,22 @@
-package kitchenpos.ui;
+package kitchenpos.table.ui;
 
+import static java.time.LocalDateTime.now;
 import static kitchenpos.util.ObjectCreator.getObject;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
-import kitchenpos.menu.service.MenuGroupService;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.dto.request.CreateMenuGroupRequest;
-import kitchenpos.menu.dto.response.MenuGroupResponse;
-import kitchenpos.menu.ui.MenuGroupRestController;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.dto.request.CreateTableGroupRequest;
+import kitchenpos.table.dto.response.TableGroupResponse;
+import kitchenpos.table.service.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,9 +25,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
 import org.springframework.test.web.servlet.MockMvc;
 
-@WebMvcTest(MenuGroupRestController.class)
+@WebMvcTest(TableGroupRestController.class)
 @MockBean(JpaMetamodelMappingContext.class)
-class MenuGroupRestControllerTest {
+class TableGroupRestControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
@@ -35,37 +36,37 @@ class MenuGroupRestControllerTest {
     private ObjectMapper objectMapper;
 
     @MockBean
-    private MenuGroupService menuGroupService;
+    private TableGroupService tableGroupService;
 
-    @DisplayName("메뉴 그룹을 생성한다")
+    @DisplayName("테이블 그룹을 생성한다")
     @Test
-    void create() throws Exception {
+    void create()
+            throws Exception {
         // given
-        final CreateMenuGroupRequest request = getObject(CreateMenuGroupRequest.class, "test");
-        final MenuGroup menuGroup = getObject(MenuGroup.class, 1L, "test");
+        final List<OrderTable> orderTables = List.of();
 
-        when(menuGroupService.create(any()))
-                .thenReturn(MenuGroupResponse.from(menuGroup));
+        final CreateTableGroupRequest request = getObject(CreateTableGroupRequest.class, 1L, orderTables);
+
+        final TableGroup tableGroup = new TableGroup(1L, now());
+
+        when(tableGroupService.create(any()))
+                .thenReturn(TableGroupResponse.from(tableGroup, orderTables));
 
         // when & then
-        mockMvc.perform(post("/api/menu-groups")
+        mockMvc.perform(post("/api/table-groups")
                         .contentType(APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(request)))
                 .andExpect(status().isCreated())
                 .andExpect(jsonPath("id").isNumber())
-                .andExpect(jsonPath("name").isString());
+                .andExpect(jsonPath("createdDate").isString())
+                .andExpect(jsonPath("orderTables").isArray());
     }
 
-    @DisplayName("메뉴 그룹 목록을 조회한다")
+    @DisplayName("테이블 그룹을 해제한다")
     @Test
-    void list() throws Exception {
-        // given
-        when(menuGroupService.list())
-                .thenReturn(List.of());
-
+    void ungroup() throws Exception {
         // when & then
-        mockMvc.perform(get("/api/menu-groups"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").isArray());
+        mockMvc.perform(delete("/api/table-groups/{tableGroupId}", 1L))
+                .andExpect(status().isNoContent());
     }
 }
